+++ src/entities/feed/model/report.type.ts
@@ -1,5 +1,12 @@
+export interface IReportFeedParamsDTO {
+  feedId: string;
+}
 export interface IReportFeedReqDTO {
   category: string;
   content: string;
   isBlind: boolean;
+}
+
+export interface IReportFeedResDTO {
+  isReported: boolean;
 }
