+++ src/entities/feed/model/write.type.ts
@@ -1,6 +1,12 @@
 import { TFeedScope } from './common.type';
 
-export interface IWriteFeedReqDTO {
+// 피드 작성
+export type TWriteFeedReqDTO = IFeedEditCommonDTO;
+
+// 피드 수정
+export type TModifyFeedReqDTO = IFeedEditCommonDTO;
+
+export interface IFeedEditCommonDTO {
   content: string;
   images: string[];
   scope: TFeedScope;
