+++ src/entities/hide/model/hide.type.ts
