+++ src/entities/hide/model/index.ts
@@ -1,1 +1,1 @@
-export * from './types';
+export * from './hide.type';
