+++ src/entities/search/model/index.ts
@@ -1,1 +1,1 @@
-export * from './types';
+export * from './search.type';
