+++ src/entities/search/model/search.type.ts
